+++ composer.json
@@ -28,6 +28,11 @@
       "Blumilk\\Codestyle\\": "src/"
     }
   },
+  "autoload-dev": {
+    "psr-4": {
+      "Blumilk\\Codestyle\\Tests\\": "tests/codestyle/"
+    }
+  },
   "scripts": {
     "cs": "./vendor/bin/php-cs-fixer fix --dry-run --diff --config codestyle.php",
     "csf": "./vendor/bin/php-cs-fixer fix --diff --config codestyle.php",
