+++ tests/codestyle/config/config.purge.php
